double preAssesment and CommentDTO instead String

+++ src/main/java/com/tyiu/corn/model/entities/Idea.java
@@ -42,7 +42,7 @@ public class Idea {
     private Long marketValue;
     private Long originality;
     private Long technicalRealizability;
-    private Long preAssessment;
+    private double preAssessment;
     private double rating;
     private double risk;
 }
